backend/requirements.txt
# ============================================
# OPSFLUX BACKEND - REQUIREMENTS PRODUCTION
# ============================================

# Django Core
Django==5.0.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5

# Base de données
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache & Queue
redis==5.0.4
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1

# API & Documentation
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
djangorestframework-simplejwt==5.3.1

# Authentification & Sécurité
django-allauth==0.61.1
python-jose[cryptography]==3.3.0
cryptography==42.0.5
argon2-cffi==23.1.0

# Fichiers & Médias
Pillow==10.3.0
python-magic==0.4.27
django-storages==1.14.2
boto3==1.34.84

# ============================================
# IA & ML - VERSIONS COMPATIBLES AVEC HTTPX 0.27
# ============================================
openai==1.23.2
anthropic==0.21.3
mistralai==0.4.2
langchain==0.1.16
langchain-community==0.0.34
tiktoken==0.6.0

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# Data Science
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# Utilitaires
python-dateutil==2.9.0
pytz==2024.1
python-decouple==3.8
requests==2.31.0

# Email
django-anymail[mailgun]==10.3

# Monitoring & Logging
sentry-sdk==1.45.0
django-extensions==3.2.3
python-json-logger==2.0.7

# WSGI Server
gunicorn==22.0.0
uvicorn[standard]==0.29.0

# MPTT
django-mptt==0.16.0

# Whitenoise
whitenoise==6.6.0
